+++ src/tightdb/table.hpp
@@ -801,8 +801,7 @@ private:
     void init(ConstSubspecRef shared_spec, ArrayParent* parent_column,
               std::size_t parent_row_ndx);
 
-    // view management support:
-    void do_erase_row(std::size_t row_ndx);
+    void do_clear();
 
     static void do_insert_column(Descriptor&, std::size_t col_ndx, DataType type,
                                  StringData name, Table* link_target_table);
@@ -1157,11 +1156,6 @@ inline void Table::bump_version(bool bump_global) const
     }
 }
 #endif
-
-inline void Table::remove(std::size_t row_ndx)
-{
-    do_erase_row(row_ndx);
-}
 
 inline void Table::remove_last()
 {