+++ release_notes.md
@@ -6,7 +6,8 @@ The C++ API has been updated and your code will break!
 
 ### Bugfixes:
 
-* None.
+* Fixed several bugs in connection to removal of like-type columns.
+* Fixed bug when last regular column is removed from table with remaining hidden backlink columns.
 
 ### API breaking changes: